backend/requirements.txt
alembic==1.15.2
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.2.0
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
Flask==3.1.1
flask-buzz==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-Mailman==1.1.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
flask-praetorian==1.6.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
passlib==1.7.4
pendulum==3.1.0
psycopg2==2.9.10
py-buzz==4.2.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
wrapt==1.17.2
